Refactor these tests
Mirror the implemenation file organisation in the tests.
